requirements.txt
pytest~=8.3.4
playwright~=1.49.0
requests
qase-pytest==4.2.0
python-dotenv~=1.0.1
Faker~=33.1.0
pydantic-settings~=2.7.0
psycopg2-binary~=2.9.10
redis~=5.2.1